refactoring
the plugin used a ToolInstallation for configuration. But logstash is
not a tool that can be installed. Instead this should be a
GlobalConfiguration.
The indexers have different configuration requirements. Host and port
are common, but things like username are not required everywhere. Also
the previous implementation was totally unflexible. In order to add a
new indexer one has to modify the sources. By configuring the indexers
via an extension point we make it easily possible to have individual
configuration jelly for each type. Adding a new indexer is just adding
new classes. So it can be even moved to a separate plugin.

Many of the fields in the plugin were public or package private. They
have been made private or protected with corresponding getters and
setters and direct calls to the fields replaced.

Get the charset of a run and use it whenever we need to convert the
String to a byte array

Use DateFormatter from apache common lang to avoid potential threading
problems.